# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dominium DayZ Server API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ServerStatus {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  serverName: string;
  map: string;
  ip: string;
  port: number;
  ping: number;
  uptime: string;
  /** @nullable */
  version?: string | null;
  password?: boolean;
}

export interface ServerStats {
  totalPlayersEver: number;
  totalHoursPlayed: number;
  totalKills: number;
  totalDeaths: number;
  totalVotes: number;
  serverAgedays: number;
}

export type NewsArticleCategory = typeof NewsArticleCategory[keyof typeof NewsArticleCategory];


export const NewsArticleCategory = {
  update: 'update',
  event: 'event',
  announcement: 'announcement',
  maintenance: 'maintenance',
} as const;

export interface NewsArticle {
  id: number;
  title: string;
  content: string;
  category: NewsArticleCategory;
  author: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  pinned?: boolean;
}

export type NewsInputCategory = typeof NewsInputCategory[keyof typeof NewsInputCategory];


export const NewsInputCategory = {
  update: 'update',
  event: 'event',
  announcement: 'announcement',
  maintenance: 'maintenance',
} as const;

export interface NewsInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  category: NewsInputCategory;
  author: string;
  imageUrl?: string;
  pinned?: boolean;
}

export type StorePackageType = typeof StorePackageType[keyof typeof StorePackageType];


export const StorePackageType = {
  vip: 'vip',
  vip_plus: 'vip_plus',
  premium: 'premium',
  item_pack: 'item_pack',
} as const;

export interface StorePackage {
  id: number;
  name: string;
  description: string;
  price: number;
  type: StorePackageType;
  features: string[];
  popular: boolean;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  badgeText?: string | null;
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  value: number;
  category: string;
  playtime: number;
  lastSeen: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type StaffMemberRole = typeof StaffMemberRole[keyof typeof StaffMemberRole];


export const StaffMemberRole = {
  owner: 'owner',
  co_owner: 'co_owner',
  admin: 'admin',
  moderator: 'moderator',
  helper: 'helper',
  developer: 'developer',
} as const;

export interface StaffMember {
  id: number;
  name: string;
  role: StaffMemberRole;
  discordTag: string;
  /** @nullable */
  avatarUrl?: string | null;
  online: boolean;
  /** @nullable */
  bio?: string | null;
}

export interface FaqItem {
  id: number;
  question: string;
  answer: string;
  category: string;
  order: number;
}

export type BanAppealStatus = typeof BanAppealStatus[keyof typeof BanAppealStatus];


export const BanAppealStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  approved: 'approved',
  denied: 'denied',
} as const;

export interface BanAppeal {
  id: number;
  playerName: string;
  steamId: string;
  reason: string;
  description: string;
  status: BanAppealStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface BanAppealInput {
  /** @minLength 1 */
  playerName: string;
  /** @minLength 1 */
  steamId: string;
  /** @minLength 1 */
  reason: string;
  /** @minLength 10 */
  description: string;
  discordTag: string;
  proofUrl?: string;
}

export type BanAppealUpdateStatus = typeof BanAppealUpdateStatus[keyof typeof BanAppealUpdateStatus];


export const BanAppealUpdateStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  approved: 'approved',
  denied: 'denied',
} as const;

export interface BanAppealUpdate {
  status?: BanAppealUpdateStatus;
  adminNote?: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  createdAt: string;
}

export interface VoteCount {
  totalVotes: number;
  monthlyVotes: number;
  weeklyVotes: number;
}

export interface VoteInput {
  /** @minLength 1 */
  playerName: string;
}

export interface OnlinePlayer {
  name: string;
  playtime: number;
  location: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface PlayerProfile {
  name: string;
  kills: number;
  deaths: number;
  playtime: number;
  lastSeen: string;
  joinedAt: string;
  kdr?: number;
  rank?: number;
  /** @nullable */
  avatarUrl?: string | null;
}

export type GetLeaderboardParams = {
category?: GetLeaderboardCategory;
limit?: number;
};

export type GetLeaderboardCategory = typeof GetLeaderboardCategory[keyof typeof GetLeaderboardCategory];


export const GetLeaderboardCategory = {
  kills: 'kills',
  survived: 'survived',
  playtime: 'playtime',
  loot: 'loot',
} as const;

export type SearchPlayerParams = {
name: string;
};