# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dominium DayZ Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /store/packages/{id}:
    get:
      operationId: getStorePackageById
      tags: [store]
      summary: Get a store package by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorePackage"
        "404":
          description: Not found
...
  /players/search:
    get:
      operationId: searchPlayer
      tags: [players]
      summary: Search for a player by name
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "404":
          description: Not found
components:
  schemas:
...
    StorePackage:
      type: object
      required: [id, name, description, price, type, features, popular]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        type:
          type: string
          enum: [vip, vip_plus, premium, item_pack]
        features:
          type: array
          items:
            type: string
        popular:
          type: boolean
        color:
          type: ["string", "null"]
        badgeText:
          type: ["string", "null"]
...
    PlayerProfile:
      type: object
      required: [name, kills, deaths, playtime, lastSeen, joinedAt]
      properties:
        name:
          type: string
        kills:
          type: integer
        deaths:
          type: integer
        playtime:
          type: integer
        lastSeen:
          type: string
        joinedAt:
          type: string
        kdr:
          type: number
        rank:
          type: integer
        avatarUrl:
          type: ["string", "null"]