# lib/api-zod/src/generated/types/playerProfile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dominium DayZ Server API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerProfile {
  name: string;
  kills: number;
  deaths: number;
  playtime: number;
  lastSeen: string;
  joinedAt: string;
  kdr?: number;
  rank?: number;
  /** @nullable */
  avatarUrl?: string | null;
}